refactor(service) moved the PasswordHasher and IDGenerator interfaces closer to where they are actually used
