--- internal/service/ports.go
@@ -1,10 +1,0 @@
-package service
-
-type PasswordHasher interface {
-	Hash(password string) (string, error)
-	Compare(hashedPassword, password string) error
-}
-
-type IDGenerator interface {
-	Generate() (string, error)
-}
